Working with tail controlled jump

+++ lib/tail_ctrl.h
@@ -2,6 +2,6 @@ void SetupTimer5();
 void setPitchControlFlag(char state);
 void tailCtrlSetup();
 void resetBodyAngle();
-void setPitchSetpoint(int setpoint);
+void setPitchSetpoint(long setpoint);
 
 extern long body_angle;
